# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Elgiganten Repair API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/"
        }
    ],
    "paths": {
...
        "/kunder/{id}/addresse": {
            "get": {
                "tags": [
                    "Personer - Kunder"
                ],
                "summary": "Find en specifik Kundes Foretrukne Adresse",
                "description": "",
                "operationId": "kunderAddressesSpecific",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Kunde ID'et",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "vej": {
                                            "type": "string",
                                            "example": "Bedste Vej Navn"
                                        },
                                        "vej_nummer": {
                                            "type": "string",
                                            "example": "12B"
                                        },
                                        "foretrukken": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "31-03-2022 02:28:47"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "31-03-2022 03:15:46"
                                        },
                                        "by": {
                                            "properties": {
                                                "postnummer": {
                                                    "type": "integer",
                                                    "maximum": 9999,
                                                    "minimum": 100,
                                                    "example": "5000"
                                                },
                                                "navn": {
                                                    "type": "string",
                                                    "example": "Odense C"
                                                }
                                            },
                                            "type": "object",
                                            "example": {
                                                "postnummer": 5000,
                                                "navn": "Odense C"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ingen Addresse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/kunder/{id}/addresser": {
            "get": {
                "tags": [
                    "Personer - Kunder"
                ],
                "summary": "Liste af en specifik Kundes Adresser",
                "description": "",
                "operationId": "kunderAddressesList",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Kunde ID'et",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "vej": {
                                            "type": "string",
                                            "example": "Bedste Vej Navn"
                                        },
                                        "vej_nummer": {
                                            "type": "string",
                                            "example": "12B"
                                        },
                                        "foretrukken": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "31-03-2022 02:28:47"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "31-03-2022 03:15:46"
                                        },
                                        "by": {
                                            "properties": {
                                                "postnummer": {
                                                    "type": "integer",
                                                    "maximum": 9999,
                                                    "minimum": 100,
                                                    "example": "5000"
                                                },
                                                "navn": {
                                                    "type": "string",
                                                    "example": "Odense C"
                                                }
                                            },
                                            "type": "object",
                                            "example": {
                                                "postnummer": 5000,
                                                "navn": "Odense C"
                                            }
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "id": 1,
                                        "vej": "Swag Vej",
                                        "vej_nummer": "12A",
                                        "foretrukken": false,
                                        "created_at": "31-03-2022 02:28:47",
                                        "updated_at": "31-03-2022 03:15:46",
                                        "by": {
                                            "postnummer": 5000,
                                            "navn": "Odense C"
                                        }
                                    },
                                    {
                                        "id": 2,
                                        "vej": "Swag Vej",
                                        "vej_nummer": "34B",
                                        "foretrukken": true,
                                        "created_at": "31-03-2022 02:28:47",
                                        "updated_at": "31-03-2022 03:15:46",
                                        "by": {
                                            "postnummer": 5000,
                                            "navn": "Odense C"
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "404": {
                        "description": "Ingen Addresser"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Personer - Kunder"
                ],
                "summary": "Opret en ny Adresse og tilknyt den specifikke Kunde",
                "description": "",
                "operationId": "kunderAddressesCreate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Kunde ID'et",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "vej",
                        "in": "query",
                        "description": "vej navnet",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "vej_nummer",
                        "in": "query",
                        "description": "vej numret",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "postnummer",
                        "in": "query",
                        "description": "postnumret",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "maximum": 9999,
                            "minimum": 100
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Addresse oprettet"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/kunder/{id}/addresser/{addresse_id}/fjern": {
            "delete": {
                "tags": [
                    "Personer - Kunder"
                ],
                "summary": "Slet en Adresse tilknytning for en specifik Kunde",
                "description": "",
                "operationId": "kunderAddressesDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Kunde ID'et",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "addresse_id",
                        "in": "path",
                        "description": "Addresse ID'et",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Addresse Slettet"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/kunder/{id}/addresser/{addresse_id}/fjern/true": {
            "delete": {
                "tags": [
                    "Personer - Kunder"
                ],
                "summary": "Slet en Adresse tilknytning for en specifik Kunde og selve addressen",
                "description": "",
                "operationId": "kunderAddressesDeleteObject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Kunde ID'et",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "addrese_id",
                        "in": "path",
                        "description": "Addresse ID'et",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Addresse Slettet"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}